Changed decimal fields to float fields on forms.py and models.py

+++ app/forms.py
@@ -1,6 +1,6 @@
 from flask_wtf import FlaskForm
 from flask_login import current_user
-from wtforms import StringField, PasswordField, SelectField, DecimalField
+from wtforms import StringField, PasswordField, SelectField, FloatField
 from wtforms.ext.dateutil.fields import DateField
 from wtforms.validators import DataRequired, Length, Email, EqualTo, ValidationError
 from app.models import User
@@ -49,13 +49,13 @@ class UpdateProfileForm(FlaskForm):
 class AddTokenForm(FlaskForm):
     tokens = SelectField('Choose Token', validators=[DataRequired()],
                          choices=[(token['symbol'], token['name']) for token in token_data['tokens']])
-    token_amount = DecimalField('Amount', validators=[DataRequired()])
-    token_price = DecimalField('Price', validators=[DataRequired()])
+    token_amount = FloatField('Amount', validators=[DataRequired()])
+    token_price = FloatField('Price', validators=[DataRequired()])
     buy_date = DateField('Date', display_format='%Y-%m-%d', validators=[DataRequired()])
 
 
 class UpdateTokenForm(FlaskForm):
     tokens = StringField('Token')
-    token_amount = DecimalField('Amount', validators=[DataRequired()])
-    token_price = DecimalField('Price', validators=[DataRequired()])
+    token_amount = FloatField('Amount', validators=[DataRequired()])
+    token_price = FloatField('Price', validators=[DataRequired()])
     buy_date = DateField('Date', display_format='%Y-%m-%d', validators=[DataRequired()])
